Move version from project to manifest
